+++ src/main/java/org/bytedeco/javacpp/FloatPointer.java
@@ -63,7 +63,7 @@ public class FloatPointer extends Pointer {
      *
      * @param size the number of {@code float} elements to allocate
      */
-    public FloatPointer(int size) {
+    public FloatPointer(long size) {
         try {
             allocateArray(size);
         } catch (UnsatisfiedLinkError e) {
@@ -74,7 +74,7 @@ public class FloatPointer extends Pointer {
     public FloatPointer() { }
     /** @see Pointer#Pointer(Pointer) */
     public FloatPointer(Pointer p) { super(p); }
-    private native void allocateArray(int size);
+    private native void allocateArray(long size);
 
     /** @see Pointer#position(long) */
     @Override public FloatPointer position(long position) {
@@ -92,7 +92,7 @@ public class FloatPointer extends Pointer {
     /** @return {@code get(0)} */
     public float get() { return get(0); }
     /** @return the i-th {@code float} value of a native array */
-    public native float get(int i);
+    public native float get(long i);
     /** @return {@code put(0, f)} */
     public FloatPointer put(float f) { return put(0, f); }
     /**
